requirements.txt
Flask==2.0.1
qrcode==7.3.1
fontTools==4.38.0
Pillow==9.4.0
spidev==3.6
unicode_slugify==0.1.5
